@@ -3,13 +3,13 @@ package com.capital7software.ai.localsearch.constraints;
 import com.capital7software.ai.localsearch.Course;
 import com.capital7software.ai.localsearch.ScheduledCourse;
 import com.capital7software.ai.localsearch.Semester;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
 import java.util.*;
 
-import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.*;
 
 public class AnyConstraintListTest {
     private static final String ICS_DEPARTMENT;
@@ -77,7 +77,7 @@ public class AnyConstraintListTest {
     private List<Constraint> semesterRestrictionList;
 
     private ScheduledCourse getScheduledCourse(String dept, int courseId) {
-        Course course = null;
+        Course course;
         ScheduledCourse scheduledCourse = null;
         int courseIndex = courseList.indexOf(new Course(dept, courseId));
 
@@ -181,7 +181,7 @@ public class AnyConstraintListTest {
         }
     }
 
-    @Before
+    @BeforeEach
     public void setupTestHarness() {
         constraintList = new AnyConstraintList();
         prerequisiteList = new ArrayList<>();
@@ -190,16 +190,16 @@ public class AnyConstraintListTest {
         loadPrerequisites();
         loadSemesterRestrictions();
 
-        for (Prerequisite prerequisite: prerequisiteList) {
+        for (Prerequisite prerequisite : prerequisiteList) {
             constraintList.add(prerequisite);
         }
 
-        for (Constraint constraint: semesterRestrictionList) {
+        for (Constraint constraint : semesterRestrictionList) {
             constraintList.add(constraint);
         }
     }
 
-    @After
+    @AfterEach
     public void tareDownTestHarness() {
         constraintList.clear();
         prerequisiteList.clear();
@@ -230,7 +230,7 @@ public class AnyConstraintListTest {
     public void getConflictsShouldBeEmpty() {
         randomlyScheduleCourses();
         int expected = 0;
-        int actual = constraintList.getConflicts().getNumberOfConflicts();;
+        int actual = constraintList.getConflicts().getNumberOfConflicts();
 
         assertEquals(expected, actual);
     }
